Add files via upload
This is a simple project where the system maintains the inventory of books in a bookshop. If a customer purchases a book, the book's count will decrease; if a book is added, the same is updated. Notice the use of pointers. You can modify the code to add a book ID and make the search based on book ID or make the search using just one parameter giving multiple results, and so on.