Add files via upload
This is an exciting project, where we will learn about the library used for random numbers: cstdlib. The program asks for a betting amount and then asks the user to guess a number on rolling. If the random number generated matches the user input, he wins, else money is deducted. The user can keep playing until he loses all the amount he put in initially.